fix package information
