+++ src/main/java/com/microsoft/jenkins/acr/QuickBuildContext.java
@@ -12,9 +12,9 @@ import com.microsoft.jenkins.acr.commands.scm.GitSCMCommand;
 import com.microsoft.jenkins.acr.commands.scm.LocalSCMCommand;
 import com.microsoft.jenkins.acr.commands.scm.TarballSCMCommand;
 import com.microsoft.jenkins.acr.common.QuickBuildRequest;
-import com.microsoft.jenkins.acr.command.scm.GitSCMRequest;
-import com.microsoft.jenkins.acr.command.scm.LocalSCMRequest;
-import com.microsoft.jenkins.acr.command.scm.RemoteTarballSCMRequest;
+import com.microsoft.jenkins.acr.common.scm.GitSCMRequest;
+import com.microsoft.jenkins.acr.common.scm.LocalSCMRequest;
+import com.microsoft.jenkins.acr.common.scm.RemoteTarballSCMRequest;
 import com.microsoft.jenkins.azurecommons.command.CommandService;
 import com.microsoft.jenkins.azurecommons.command.IBaseCommandData;
 import com.microsoft.jenkins.azurecommons.command.ICommand;
