+++ src/main/java/com/microsoft/jenkins/acr/commands/scm/package-info.java
@@ -4,10 +4,10 @@
  */
 
 /**
- * SCM management package.
+ * Command to resolve the build request source url.
  *
  * @since 1.0
  * @author yuwzho
  * @version 1.0
  */
-package com.microsoft.jenkins.acr.command.scm;
+package com.microsoft.jenkins.acr.commands.scm;
