+++ src/main/java/com/microsoft/jenkins/acr/common/scm/LocalSCMRequest.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+package com.microsoft.jenkins.acr.common.scm;
+
+public interface LocalSCMRequest {
+    String getLocalDir();
+}
