+++ src/main/java/com/microsoft/jenkins/acr/common/scm/RemoteTarballSCMRequest.java
@@ -3,8 +3,8 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.jenkins.acr.command.scm;
+package com.microsoft.jenkins.acr.common.scm;
 
-public interface LocalSCMRequest {
-    String getLocalDir();
+public interface RemoteTarballSCMRequest {
+    String getTarball();
 }
