+++ src/main/java/com/microsoft/jenkins/acr/common/scm/SCMRequest.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.jenkins.acr.command.scm;
+package com.microsoft.jenkins.acr.common.scm;
 
 public interface SCMRequest extends GitSCMRequest, LocalSCMRequest, RemoteTarballSCMRequest {
